+++ src/com/csc422/zombiewar/human/Child.java
@@ -1,4 +1,4 @@
-package com.csc422.zombiewar;
+package com.csc422.zombiewar.human;
 
 public class Child extends Survivor {
 	public Child() {
