+++ src/com/csc422/zombiewar/human/CommonInfected.java
@@ -1,4 +1,4 @@
-package com.csc422.zombiewar;
+package com.csc422.zombiewar.human;
 
 public class CommonInfected extends Zombie {
 	public CommonInfected() {
