+++ src/com/csc422/zombiewar/human/Soldier.java
@@ -1,4 +1,4 @@
-package com.csc422.zombiewar;
+package com.csc422.zombiewar.human;
 
 public class Soldier extends Survivor {
 	public Soldier() {
